css files moved to each individual component
