+++ frontend/src/components/shared/SearchBar.css
