+++ frontend/src/main.jsx
@@ -2,7 +2,6 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { App } from './App';
 import './index.css';
-import './styles/App.css';
 
 
 createRoot(document.getElementById('root')).render(
