--- frontend/src/styles/App.css
@@ -1,76 +1,0 @@
-.app {
-  min-height: 100vh;
-  background-color: #dfdfdf;
-}
-
-.main-content {
-  max-width: 1200px;
-  margin: 0 auto;
-  padding: 2rem 1rem;
-}
-
-.role-list {
-  background: white;
-  border-radius: 8px;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-  padding: 2rem;
-}
-
-.role-list h1 {
-  font-size: 1.5rem;
-  font-weight: 600;
-  color: #1f2937;
-  margin-bottom: 1.5rem;
-}
-
-.table-container {
-  overflow-x: auto;
-}
-
-table {
-  width: 100%;
-  border-collapse: collapse;
-}
-
-th {
-  text-align: left;
-  padding: 1rem;
-  background-color: #f9fafb;
-  color: #4b5563;
-  font-weight: 600;
-  border-bottom: 2px solid #e5e7eb;
-}
-
-td {
-  padding: 1rem;
-  border-bottom: 1px solid #e5e7eb;
-  color: #1f2937;
-}
-
-select {
-  padding: 0.5rem;
-  border: 1px solid #d1d5db;
-  border-radius: 4px;
-  background-color: white;
-  color: #1f2937;
-  cursor: pointer;
-}
-
-select:hover {
-  border-color: #9ca3af;
-}
-
-select:focus {
-  outline: none;
-  border-color: #3b82f6;
-  ring: 2px solid #bfdbfe;
-}
-
-.error-message {
-  background-color: #f8d7da;
-  color: #721c24;
-  padding: 10px;
-  border: 1px solid #f5c6cb;
-  margin-bottom: 20px;
-  border-radius: 4px;
-}
